Further code fixes and Visual Upgrades

+++ Assets/Scripts/Extra Scripts/MatchData.cs
@@ -6,6 +6,6 @@ public class MatchData
 	public enum Side{Izq, Der}
 	public static Side WinnerSide;
 
-	public static int WinnerPoints;
-	public static int LoserPoints;
+	public static float WinnerPoints;
+	public static float LoserPoints;
 }
